moved selectors into a selectors folder
